Upgrade to GNOME 3.15 unstable release
GNOME 3.14 is already available in the official FreeBSD ports! This is also
the reason why I don't make new builds for about two months. As installing
GNOME 3 from the official pkgng repo is easy now, I switched to the development
branch to get different things in this LiveCD. It may be more exciting to have
a JHBuild-based LiveCD, but I still don't know how to make GDM work in a
JHBuild prefix.