Add Google Noto fonts and fix fontconfig cache
In order to allow users to switch to languages other than English, we
have to include fonts on the LiveCD. GNOME control center hides
languages with no font installed when displaying the list of languages.
Even if users set locale variables manually, the user interface is not
usable because of missing glyphs. Therefore, Google Noto fonts is added
to the list of packages. It takes a lot of space, but I think having a
good-looking font is worth it.

However, the addition of noto triggers a bug in scripts used to build
LiveCD. fontconfig records inode numbers in its cache files, and it will
try to regenerate the cache when inode numbers don't match. Since the
UFS-formatted system image is created by makefs, the inode numbers of
font files are changed when they are copied from the host filesystem to
the image file, which causes fontconfig to regenerate the cache on every
boot. This was not a big problem because generating caches was very fast
and nobody seems to discover it. It become a real problem after noto is
added because it takes several minutes to generate caches for these
large sets of fonts, causing GDM to hang and the memory disk mounted for
writing to run out of space quickly.

To avoid generating font caches on boot, we have to run 'fc-cache' on
the system image itself. The previous 'uzip' step is now split into
three steps, 'ufs', 'font', 'uzip'. The 'font' step mounts the system
image created in the 'ufs' step, deletes and regenerates fontconfig
caches in its /var/db/fontconfig. Therefore, fontconfig cache files are
considered valid in the LiveCD environment, and GDM can startup quickly.